#5 Updating dependencies and taking server-test-s…
…upport into use.
