+++ .eslintrc.output.cjs
@@ -1,11 +1,14 @@
 module.exports = {
   root: true,
   extends: [
-    "eslint:recommended",
     // See https://github.com/prettier/eslint-config-prettier/blob/main/CHANGELOG.md#version-800-2021-02-21
     "plugin:prettier/recommended",
   ],
   plugins: ["prettier"],
+  parser: "@babel/eslint-parser",
+  parserOptions: {
+    requireConfigFile: false
+  },
   env: {
     node: true,
     es2020: true
