+++ .github/workflows/build-and-test.yml
@@ -1,4 +1,4 @@
-name: Build, test, and publish code coverage
+name: Build, test, and publish
 
 on:
   workflow_call:
@@ -86,7 +86,8 @@ jobs:
           cat '${{ matrix.dir }}/package.json' | jq -M 'del(.type)' > '${{ matrix.dir }}/package.json.new'
           mv '${{ matrix.dir }}/package.json.new' '${{ matrix.dir }}/package.json'
           ./scripts/build.sh '${{ matrix.dir }}' ci
-          ls -al '${{ matrix.dir }}/dist'
+          ls -al '${{ matrix.dir }}/dist-cjs' || echo 'No dist folder - TS-only package!'
+          ls -al '${{ matrix.dir }}/dist-mjs' || echo 'No dist folder - TS-only package!'
 
       - id: coverage
         name: Upload coverage for '${{ matrix.dir }}'
