+++ .gitignore
@@ -1,7 +1,7 @@
 node_modules
 .vscode
 build
-dist
+dist*
 coverage
 .yarn
 */.eslintrc*.cjs
