--- server/src/__test__/context.ts
@@ -1,15 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import type * as ctx from "../context-types";
-
-export const dummyContext: ctx.Context<State> = {
-  res: {
-    locals: {
-      property: "Property",
-    },
-  },
-} as any;
-
-export interface State {
-  property: string;
-}
