+++ server/src/context.d.ts
@@ -1,4 +1,9 @@
+import * as server from "@ty-ras/server";
 import type * as express from "express";
+
+export interface HKTContext extends server.HKTContext {
+  readonly type: Context<this["_TState"]>;
+}
 
 export type Context<T> = {
   req: express.Request;
